Building the page from redis
Work in progress

The data is returned together with the page and rendering is done in
javascript.

Still to tie back interaction with the supervisors, visibility controls.
However the page is no more organized rigidly in a
supervisor/group/process hierarchy but it can be grouped according to
different process properties and tags.